way better type system – baseGroup.ts for panel/tab/group
