+++ src/lib/gui/components/groups/index.ts
@@ -1,0 +1,5 @@
+import { Panel } from './panel';
+import { Group } from './group';
+import { Tab } from './tab';
+
+export { Panel, Group, Tab };
